requirements.txt
allure-pytest==2.13.5
grpcio-tools==1.70.0
grpcio==1.70.0
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3